Revamp the syntax tree
* The tree structure is nice and named, but it also gets rather
unergonomic soon. I therefore decided to switch to a vector based
representation.

* While working on this I realised that `:variant` in Akar is useless
in practice. I think I am going to cull it. (The joys of having a
library nobody is using!)

* I made `define`, `if` into proper AST nodes, rather than doing that
"parsing" in interpretation phase.

* This also means When parsing `(define x 11 1 1 11)` fails, it will be
parsed as a Lox list. To prevent that I disallowed specific tokens as
symbol names. This is not the best solution but I have decided not to
care about it for now.

* Writing parsers, even using combinator libraries, is hard. It took me
too long to perform this change. It wasn't easy to reason about
behaviour of many combinators. And even now, it's holding together
after some random trials and errors.

* I changed `set` to `asssign`.