Do not treat native functions as regular Lox values
Earlier we took a shortcut wherein Clojure's `+` etc were put into
global scope, and invoked directly. Here we were piggybacking directly
onto Clojure's IFn.

As of this commit, native functions are treated somewhat specially.
Our symbol resolution first tries to resolve the variable locally, and
on failure, tries to resolve it as a native function. We might add more
things to this ladder later. This is how the resolution works in actual
Lisps too.

This also paves way for addition of other Lox callables, such as
Lox functions, methods.